requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
pytest-asyncio==1.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
